Changed indentation :-P. Removed some stupid casts.

+++ event.c
@@ -36,91 +36,93 @@ int kq = -1;
 
 void init_events(void)
 {
-  struct file *file;
-  struct kevent *kevlist, *kevptr;
-  int kevlen;
+        struct file *file;
+        struct kevent *kevlist, *kevptr;
+        int kevlen;
 
-  if(kq == -1)
-  {
-    kq = kqueue();
-    if(kq == -1)
-      die("kqueue: %s", strerror(errno));
-  }
+        if(kq == -1)
+        {
+                kq = kqueue();
+                if(kq == -1)
+                        die("kqueue: %s", strerror(errno));
+        }
 
-  kevlen = count_open_files() * 2;
-  if(kevlen == 0)
-    return;
+        kevlen = count_open_files() * 2;
+        if(kevlen == 0)
+                return;
 
-  kevlist = xmalloc(sizeof(struct kevent) * kevlen);
+        kevlist = xmalloc(sizeof(struct kevent) * kevlen);
 
-  kevptr = kevlist;
-  for(file = files.head; file != NULL; file = file->next)
-  {
-    if(file->fd != NULL)
-    {
-      EV_SET(kevptr, fileno(file->fd), EVFILT_VNODE, EV_ADD | EV_CLEAR, NOTE_RENAME | NOTE_DELETE, 0, NULL);
-      kevptr++;
-      EV_SET(kevptr, fileno(file->fd), EVFILT_READ, EV_ADD | EV_CLEAR, 0, 0, NULL);
-      kevptr++;
-    }
-  }
+        kevptr = kevlist;
+        for(file = files.head; file != NULL; file = file->next)
+        {
+                if(file->fd != NULL)
+                {
+                        EV_SET(kevptr, fileno(file->fd), EVFILT_VNODE, EV_ADD |
+                            EV_CLEAR, NOTE_RENAME | NOTE_DELETE, 0, NULL);
+                        kevptr++;
+                        EV_SET(kevptr, fileno(file->fd), EVFILT_READ, EV_ADD |
+                            EV_CLEAR, 0, 0, NULL);
+                        kevptr++;
+                }
+        }
 
-  if(kevent(kq, kevlist, kevlen, NULL, 0, NULL))
-    die("kevent: %s", strerror(errno));
+        if(kevent(kq, kevlist, kevlen, NULL, 0, NULL))
+                die("kevent: %s", strerror(errno));
 
-  free(kevlist);
+        free(kevlist);
 }
 
 struct file *get_event(enum event *event, int timeout)
 {
-  struct file *file;
-  struct kevent kev;
-  struct timespec ts;
-  int rc;
+        struct file *file;
+        struct kevent kev;
+        struct timespec ts;
+        int rc;
 
-  *event = EVENT_NONE;
+        *event = EVENT_NONE;
 
-  if(kq == -1)
-    return NULL;
+        if(kq == -1)
+                return NULL;
 
-  ts.tv_nsec = 0;
-  ts.tv_sec = timeout;
+        ts.tv_nsec = 0;
+        ts.tv_sec = timeout;
 
-  rc = kevent(kq, NULL, 0, &kev, 1, &ts);
+        rc = kevent(kq, NULL, 0, &kev, 1, &ts);
 
-  if(rc == -1)
-  {
-    if(errno == EINTR) /* && !debug) */
-      return NULL;
+        if(rc == -1)
+        {
+                if(errno == EINTR) /* && !debug) */
+                        return NULL;
 
-    die("kevent: %s", strerror(errno));
-  }
+                die("kevent: %s", strerror(errno));
+        }
 
-  if(rc == 0)
-  {
-    *event = EVENT_TIMEOUT;
-    return NULL;
-  }
+        if(rc == 0)
+        {
+                *event = EVENT_TIMEOUT;
+                return NULL;
+        }
 
-  file = find_file_by_fd((int) kev.ident);
-  if(file == NULL)
-    return NULL;
+        file = find_file_by_fd(kev.ident);
+        if(file == NULL)
+                return NULL;
 
-  switch(kev.filter)
-  {
-    case EVFILT_VNODE:
-      *event = EVENT_REOPEN;
-      return file;
-    case EVFILT_READ:
-      if(kev.data < 0)
-	*event = EVENT_REOPEN;
-      else
-      {
-	file->size += kev.data;
-	*event = EVENT_READ;
-      }
-      return file;
-  }
+        switch(kev.filter)
+        {
+        case EVFILT_VNODE:
+                *event = EVENT_REOPEN;
+                return file;
+        case EVFILT_READ:
+                if(kev.data < 0)
+                        *event = EVENT_REOPEN;
+                else
+                {
+                        file->size += kev.data;
+                        *event = EVENT_READ;
+                }
+                return file;
+        }
 
-  return NULL;
+        return NULL;
 }
